api/requirements_dev.txt
pytest==6.0.1
black==19.10b0
mypy==0.782
diagrams==0.16.0
flake8==3.8.3
boto3-stubs==1.14.36.0
